Implemented the remaining features (i.e. MultiTouchChanged and FrameChanged event) of MicrosoftSurfaceTouchInputProvider for WPF 4.0 applications.

+++ Src/Silverlight/Gestures/Objects/FrameInfo.cs
@@ -27,7 +27,7 @@ namespace Gestures.Objects
             }
         }
 
-        public int TimeStamp { get; set; }
+        public long TimeStamp { get; set; }
         public int WaitTime { get; set; }
     }
 }
